Create Smart fish feeding system
Smart fish feeding aims to minimize food waste and increase the food conversion ratio (FCR) by providing an optimizing diet for fish. Smart fish feeding method with manual feeders face challenges like the use of antibiotics, risk of feed poisoning and competition among the fish. The issues of animal health like unusual physical or mental characteristics such as a change in the skin or appearance, most importantly wasted feed, and also the use of antibiotics in the filtration system, and secretions of nutrients is not perfect but it still has some impact on aquaculture. An adaptive smart fish feeding based on fish behaviour is proposed to mitigate the impact of conventional feeding mechanisms. The robot fish feeder monitors the activity of the fish before and after the feeding. A theoretical smart fish feeder attempts to give its users exactly what they want depending on the set time. This research study expect the fish to produce a standardized fish trait inside by adjusting the individual response to the requirements of the fish. This system reduces feed wastage with a number of user involvement, and operational costs. Moreover, a smart software algorithm is proposed for.